pyproject.toml
[tool.poetry]
name = "documentation-helper"
version = "0.1.0"
description = ""
authors = ["swhite92163 <stephenericwhite@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "3.9.4"
aiosignal = "1.3.1"
altair = "5.3.0"
annotated-types = "0.6.0"
anyio = "4.3.0"
attrs = "23.2.0"
backoff = "2.2.1"
beautifulsoup4 = "4.12.3"
black = "24.4.0"
blinker = "1.7.0"
cachetools = "5.3.3"
certifi = "2024.2.2"
chardet = "5.2.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
dataclasses-json = "0.6.4"
dataclasses-json-speakeasy = "0.5.11"
deepdiff = "7.0.1"
distro = "1.9.0"
emoji = "2.11.0"
fastapi = "0.110.1"
filetype = "1.2.0"
frozenlist = "1.4.1"
gitdb = "4.0.11"
gitpython = "3.1.43"
greenlet = "3.0.3"
h11 = "0.14.0"
httpcore = "1.0.5"
httpx = "0.27.0"
idna = "3.7"
jinja2 = "3.1.3"
joblib = "1.4.0"
jsonpatch = "1.33"
jsonpath-python = "1.0.6"
jsonpointer = "2.4"
jsonschema = "4.21.1"
jsonschema-specifications = "2023.12.1"
langchain = "0.1.16"
langchain-community = "0.0.32"
langchain-core = "0.1.42"
langchain-openai = "0.1.3"
langchain-text-splitters = "0.0.1"
langdetect = "1.0.9"
langsmith = "0.1.47"
lxml = "5.2.1"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
marshmallow = "3.21.1"
mdurl = "0.1.2"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
nltk = "3.8.1"
numpy = "1.26.4"
openai = "1.17.1"
ordered-set = "4.1.0"
orjson = "3.10.0"
packaging = "23.2"
pandas = "2.2.2"
pathspec = "0.12.1"
pillow = "10.3.0"
pinecone-client = "3.2.2"
platformdirs = "4.2.0"
protobuf = "4.25.3"
pyarrow = "15.0.2"
pydantic = "2.7.0"
pydantic-core = "2.18.1"
pydeck = "0.8.1b0"
pygments = "2.17.2"
pypdf = "4.2.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
python-iso639 = "2024.2.7"
python-magic = "0.4.27"
pytz = "2024.1"
pyyaml = "6.0.1"
rapidfuzz = "3.8.1"
referencing = "0.34.0"
regex = "2023.12.25"
requests = "2.31.0"
rich = "13.7.1"
rpds-py = "0.18.0"
six = "1.16.0"
smmap = "5.0.1"
sniffio = "1.3.1"
soupsieve = "2.5"
sqlalchemy = "2.0.29"
starlette = "0.37.2"
streamlit = "1.33.0"
streamlit-chat = "0.1.1"
tabulate = "0.9.0"
tenacity = "8.2.3"
tiktoken = "0.6.0"
toml = "0.10.2"
toolz = "0.12.1"
tornado = "6.4"
tqdm = "4.66.2"
typing-inspect = "0.9.0"
typing-extensions = "4.11.0"
tzdata = "2024.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
